gh-994
Review comments
